requirements.txt
Django==4.1.7
python-dotenv==0.21.1
python-telegram-bot==13.14
requests==2.28.2
